alongTrackDistance(geosphere)	R Documentation

Along Track Distance

Description

The along track distance is the distance from the start point (p1) to the closest point on the path to a third point (p3), following a great circle path defined by points p1 and p2

Usage


alongTrackDistance(p1, p2, p3, r=6378137)

Arguments

p1 	 longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 as above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

p3 	 as above

r 	 radius of the earth; default = 6378137m

Value

A distance in units of r (default is meters)

Author(s)

Chris Veness and Robert Hijmans

See Also

alongTrackDistance

Examples


alongTrackDistance(c(0,0),c(90,90),c(80,80))

[Package geosphere version 0.2 Index]