antipode(geosphere)	R Documentation

Antipodes

Description

Compute an antipode, or check whether two points are antipodes. Antipodes are places on Earth that are diametrically opposite to one another; and could be connected by a straight line through the centre of the Earth.

Antipodal points are connected by an infinite number of great circles (e.g. the meridians connecting the poles), and can therefore not be used in some great circle based computations.

Usage


antipode(p)
isAntipodal(p1, p2)

Arguments

p 	 Longitude/latitude of a single point; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p1 	 as above.

p2 	 as above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

Value

antipodal points or a logical value

Author(s)

Robert Hijmans

References

http://en.wikipedia.org/wiki/Antipodes

Examples


antipode(rbind(c(5,52), c(-120,37), c(-60,0), c(0,70)))
isAntipodal(c(0,0), c(180,0))

[Package geosphere version 0.2 Index]