bearing(geosphere)	R Documentation

Bearing

Description

Get the initial bearing (direction to travel in) to go from point1 to point2 following the shortest path (a great circle). Note that bearings change continuously while traveling along a great circle. A route with constant bearing is a rhumb line (see brngRhumb.

Usage


bearing(p1, p2)

Arguments

p1 	 longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 as above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

Value

A bearing in degrees

Author(s)

Chris Veness; ported to R by Robert Hijmans;

References

http://www.movable-type.co.uk/scripts/latlong.html

http://williams.best.vwh.net/ftp/avsig/avform.txt

See Also

brngRhumb

Examples


bearing(c(0,0),c(90,90))

[Package geosphere version 0.2 Index]