brngRhumb(geosphere)	R Documentation

Rhumbline bearing

Description

Bearing (direction of travel) along a rhumb line. Unlike a great circle, a rhumb line is a line of constant bearing.

Usage


brngRhumb(p1, p2)

Arguments

p1 	 longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 as above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

Value

A bearing in degrees

Author(s)

Chris Veness; ported to R by Robert Hijmans

References

http://www.movable-type.co.uk/scripts/latlong.html

See Also

bearing, distRhumb

Examples


brngRhumb(c(0,0),c(90,90))

[Package geosphere version 0.2 Index]