crossingParallels(geosphere)	R Documentation

Crossing parellels

Description

Longitudes at which a given great circle crosses a given parallel (latitude)

Usage


crossingParallels(p1, p2, lat) 

Arguments

p1 	 longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 as above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

lat 	 a latitude

Value

two points (longitudes)

Author(s)

Robert Hijmans based on code by Ed Williams

References

http://williams.best.vwh.net/avform.htm#Intersection

Examples


crossingParallels(c(5,52), c(-120,37), 40)

[Package geosphere version 0.2 Index]