destPointRhumb(geosphere)	R Documentation

Destination along a rhumb line

Description

Calculate the destination point when travelling along a 'rhumb line' (loxodrome), given a start point, bearing, and distance.

Usage


destPointRhumb(p, brng, dist, r = 6378137)

Arguments

p 	 longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

brng 	 bearing in degrees

dist 	 distance; in the same unit as r (default is meters)

r 	 radius of the earth; default = 6378137 m

Value

Coordinates (longitude/latitude) of a point

Author(s)

Chris Veness; ported to R by Robert Hijmans

References

http://www.movable-type.co.uk/scripts/latlong.html

See Also

destPoint

Examples


destPointRhumb(c(0,0), 30, 100000, r = 6378137)

[Package geosphere version 0.2 Index]