greatCircle(geosphere)	R Documentation

Intersecting radials

Description

Get points on a great circle as defined by the shortest distance between two specified points

Usage


greatCircle(p1, p2, n=360) 

Arguments

p1 	 Longitude/latitude of a single point; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 As above

n 	 The requested number of points on the Great Circle

Value

a matrix of points

Author(s)

Robert Hijmans based on a formula by Ed Williams

References

http://williams.best.vwh.net/avform.htm#Int

Examples


greatCircle(c(5,52), c(-120,37), n=36)

[Package geosphere version 0.2 Index]