greatCircleIntermediat(geosphere)	R Documentation

Intermediate points on a great circle

Description

Get intermediate points on a great circle inbetween the two points used to define the circle

Usage


greatCircleIntermediate(p1, p2, n=50) 

Arguments

p1 	 Longitude/latitude of a single point; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 As above

n 	 The requested number of points on the Great Circle

Value

a matrix of points

Author(s)

Robert Hijmans based on code by Ed Williams

References

http://williams.best.vwh.net/avform.htm#Intermediate

Examples


greatCircleIntermediate(c(5,52), c(-120,37), n=10)

[Package geosphere version 0.2 Index]