greatCircleIntersect(geosphere)	R Documentation

Intersections of two great circles

Description

Get the two points where two great cricles cross each other. Great circles are defined by two points on it.

Usage


greatCircleIntersect(p1, p2, p3, p4) 

Arguments

p1 	 Longitude/latitude of a single point; can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p2 	 As above

p3 	 As above

p4 	 As above

Value

two points for each pair of great circles

Author(s)

Robert Hijmans, based on equations by Ed Williams (see reference)

References

http://williams.best.vwh.net/intersect.htm

Examples


p1 <- c(5,52); p2 <- c(-120,37); p3 <- c(-60,0); p4 <- c(0,70)
greatCircleIntersect(p1,p2,p3,p4)

[Package geosphere version 0.2 Index]