polePoint(geosphere)	R Documentation

Highest latitude on a great circle

Description

Given a latitude and an initial bearing, what is the polar-most point that will be reached when following a great circle? Computed with Clairaut's formula.

Usage


polePoint(lat, brng)

Arguments

lat 	 latitude of point(s)

brng 	 bearing

Value

A pair of coordinates (longitude/latitude)

Author(s)

Chris Veness; ported to R by Robert Hijmans

References

http://williams.best.vwh.net/ftp/avsig/avform.txt

http://www.movable-type.co.uk/scripts/latlong.html

Examples


polePoint(c(5,52),30)

[Package geosphere version 0.2 Index]