radialIntersect(geosphere)	R Documentation

Intersecting radials

Description

Intersection between two lines defined by their point of origin and true bearing

Usage


radialIntersect(p1, brng1, p2, brng2) 

Arguments

p1 	 Longitude/latitude of point(s); can be a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

brng1 	 True bearing from p1

p2 	 As above. Should have same length as p1, or a single point (or vice versa when p1 is a single point

brng2 	 True bearing from p2

Value

a point

Author(s)

Robert Hijmans based on code by Ed Williams

References

http://williams.best.vwh.net/avform.htm#Intersection

Examples


radialIntersect(c(10,0), 10, c(-10,0), 10)

[Package geosphere version 0.2 Index]