utilities(geosphere)	R Documentation

Internal utility functions

Description

pointsToMatrix is a simple helper function. Point input to all other functions is processed by this function to coerce points into a two column matrix of longitude / latitude.

compareDim compares two or three points objects to check if they have the same length or have lenght 1 (point).

Usage


pointsToMatrix(p)
compareDim(p1,p2,p3)

Arguments

p 	 a vector of two numbers, a matrix of 2 columns (first one is longitude, second is latitude) or a spatialPoints* object

p1 	 a matrix of 2 columns

p2 	 a matrix of 2 columns

p3 	 a matrix of 2 columns

Value

pointsToMatrix: a matrix with two columns representing longitude and latitude.
compareDim: TRUE (invisible) or an error

Author(s)

Robert Hijmans and Jacob van Etten

[Package geosphere version 0.2 Index]